fix: split worker.go to satisfy file-length lint
